+++ popolo/models.py
@@ -513,7 +513,37 @@ class IdentifierShortcutsMixin(object):
 
 class ClassificationShortcutsMixin(object):
 
-    def add_classification(self, classification, **kwargs):
+    def add_classification(self, scheme, code=None, descr=None, **kwargs):
+        """Add classification to the instance inheriting the mixin
+        :param scheme: classification scheme (ATECO, LEGAL_FORM_IPA, ...)
+        :param code:   classification code, internal to the scheme
+        :param descr:  classification textual description (brief)
+        :param kwargs: other params as source, start_date, end_date, ...
+        :return: the classification instance just added
+        """
+        # classifications having the same scheme, code and descr are considered
+        # overlapping and will not be added
+        if code is None and descr is None:
+            raise Exception(
+                "At least one between descr "
+                "and code must take value"
+            )
+
+        # first create the Classification object,
+        # or fetch an already existing one
+        c, created = Classification.objects.get_or_create(
+            scheme=scheme,
+            code=code,
+            descr=descr,
+            defaults=kwargs
+        )
+
+        # then add the ClassificationRel to classifications
+        self.classifications.get_or_create(
+            classification=c
+        )
+
+    def add_classification_rel(self, classification, **kwargs):
         """Add classification (rel) to the instance inheriting the mixin
 
         :param classification_id: existing Classification id
@@ -536,7 +566,11 @@ class ClassificationShortcutsMixin(object):
         """
         # add objects
         for new_classification in new_classifications:
-            self.add_classification(**new_classification)
+            if 'classification' in new_classification:
+                self.add_classification_rel(**new_classification)
+            else:
+                self.add_classification(**new_classification)
+
 
     def update_classifications(self, new_classifications):
         """update classifications,
@@ -1289,12 +1323,19 @@ class Organization(
         _("classification"),
         max_length=256,
         blank=True, null=True,
-        help_text=_("The main classification, legal form in many cases")
+        help_text=_("The nature of the organization, legal form in many cases")
+    )
+
+    thematic_classification = models.CharField(
+        _("thematic classification"),
+        max_length=256,
+        blank=True, null=True,
+        help_text=_("What the organization does, in what fields, ...")
     )
 
     classifications = GenericRelation(
         'ClassificationRel',
-        help_text=_("ATECO, Legal Form and other classifications")
+        help_text=_("ATECO, Legal Form and all other available classifications")
     )
 
     # array of items referencing